pyproject.toml
[project]
name = "mcp-server-azure-ai-search-preview"
version = "0.3.1"
description = "An MCP Service for Azure AI Search"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-search-documents>=11.5.2",
    "azure-storage-blob>=12.25.1",
    "hatchling>=1.27.0",
    "mcp==1.6.0",
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server_azure_ai_search_preview"]
